Speed up excel exports of repositories.
Fetching local_roles only once and cache them.
