+++ changes/CA-5643.other
@@ -1,0 +1,1 @@
+Speed up excel exports of repositories. [phgross]
